+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Added `craft\shopify\elements\Product::getDefaultVariant()`.
+- Added `craft\shopify\elements\Product::getCheapestVariant()`.
 - Fixed a PHP error that occurred when saving invalid settings. ([#39](https://github.com/craftcms/shopify/pull/39),[#40](https://github.com/craftcms/shopify/pull/40))
 
 ## 3.0.0.1 - 2022-11-08